requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pyyaml==6.0.2

# Dashboard
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3

# Dev & testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Optional: ML classifier
scikit-learn==1.5.2
numpy==2.1.2
